+++ EstiPro.Api.sln
@@ -17,8 +17,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EstiPro.Application", "src\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EstiPro.Infrastructure", "src\EstiPro.Infrastructure\EstiPro.Infrastructure.csproj", "{C8B450AC-8AC8-46A5-BEA0-D8A932A144FA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EstiPro.Presentation", "src\EstiPro.Presentation\EstiPro.Presentation.csproj", "{F64C8863-3C75-4403-AC74-2B13757590F9}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EstiPro.Api", "src\EstiPro.Api\EstiPro.Api.csproj", "{A1248C42-3BB4-4A2D-9B84-617B2C32F6F5}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Files", "Solution Files", "{46A667D6-E533-428A-A5FE-75A4C29963F2}"
@@ -57,10 +55,6 @@ Global
 		{C8B450AC-8AC8-46A5-BEA0-D8A932A144FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C8B450AC-8AC8-46A5-BEA0-D8A932A144FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C8B450AC-8AC8-46A5-BEA0-D8A932A144FA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F64C8863-3C75-4403-AC74-2B13757590F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F64C8863-3C75-4403-AC74-2B13757590F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F64C8863-3C75-4403-AC74-2B13757590F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F64C8863-3C75-4403-AC74-2B13757590F9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A1248C42-3BB4-4A2D-9B84-617B2C32F6F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A1248C42-3BB4-4A2D-9B84-617B2C32F6F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1248C42-3BB4-4A2D-9B84-617B2C32F6F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -79,6 +73,5 @@ Global
 		{AA0AD642-B49A-4C8B-970F-9E868261A283} = {79ED024F-BA53-482D-B7D4-6162B2BF19DA}
 		{62BF4E42-57FA-4DDF-9860-6DC7329A34E7} = {79ED024F-BA53-482D-B7D4-6162B2BF19DA}
 		{C8B450AC-8AC8-46A5-BEA0-D8A932A144FA} = {79ED024F-BA53-482D-B7D4-6162B2BF19DA}
-		{F64C8863-3C75-4403-AC74-2B13757590F9} = {79ED024F-BA53-482D-B7D4-6162B2BF19DA}
 	EndGlobalSection
 EndGlobal
