--- src/EstiPro.Presentation/EstiPro.Presentation.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <OutputType>Library</OutputType>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="FluentResults.Extensions.AspNetCore"/>
-        <PackageReference Include="MediatR"/>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi"/>
-        <PackageReference Include="Newtonsoft.Json"/>
-        <PackageReference Include="Swashbuckle.AspNetCore"/>
-        <PackageReference Include="Swashbuckle.AspNetCore.Annotations"/>
-        <PackageReference Include="System.Text.Encodings.Web"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\EstiPro.Application\EstiPro.Application.csproj"/>
-        <ProjectReference Include="..\EstiPro.Domain\EstiPro.Domain.csproj"/>
-    </ItemGroup>
-
-</Project>
